refactor(test): extract FakeD1 harness to worker/test/fake-d1.ts (#159)
Core makeFakeStmt/makeFakeDb/capture helpers live outside src/.
test-utils.ts re-exports for backward-compatible imports.
